Slot_attention/claude code/requirements.txt
# PyTorch 관련 패키지
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# 데이터 처리
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# 이미지 처리
Pillow>=8.3.0
opencv-python>=4.5.0

# 시각화
matplotlib>=3.5.0
seaborn>=0.11.0

# 머신러닝 유틸리티
tqdm>=4.62.0
tensorboard>=2.7.0

# 선택적 패키지 (고급 기능용)
wandb>=0.12.0  # 실험 추적용 (선택사항)
h5py>=3.6.0    # 데이터 저장용 (선택사항)

# 개발/테스트용
pytest>=6.2.0
jupyter>=1.0.0
ipython>=7.0.0